rename and orginaized tutorial resource folder
