+++ tutorials/mct_model_garden/evaluation_metrics/coco_evaluation.py
