+++ tutorials/mct_model_garden/models_keras/yolov8/yolov8.py
