--- tutorials/mct_models_lib/README.md
@@ -1,18 +1,0 @@
-# MCT Models Library 
-
-MCT Models Library is a collection of models sourced from various of repositories and adjusted for Model Compression Toolkit (MCT).
-
-## Models
-
-| Model        | Source Repository           | MCT Model Library                                                                                                     | 
-|--------------|-----------------------------|-----------------------------------------------------------------------------------------------------------------------|
-| EfficientDet | [efficientdet-pytorch](https://github.com/rwightman/efficientdet-pytorch) | [model code](https://github.com/sony/model_optimization/tree/main/tutorials/mct_models_lib/models_keras/efficientdet) |
-| Nanodet-Plus | [Nanodet-Plus](https://github.com/RangiLyu/nanodet)         | [model code](https://github.com/sony/model_optimization/tree/main/tutorials/mct_models_lib/models_keras/nanodet)      |                                                                                                                
-| Yolov8n      | [Ultralytics](https://github.com/ultralytics/ultralytics)          | [model code](https://github.com/sony/model_optimization/tree/main/tutorials/mct_models_lib/models_keras/yolov8)       |
-
-
-
-## License
-This project is licensed under [Apache License 2.0](../../LICENSE.md).
-However, please note that different repositories have their own licenses. Therefore, when using a model from 
-this library, it's essential to also comply with the licensing terms of the source repositories.
